similarity-g3/requirements.txt
deepface==0.0.92
tf-keras==2.16.0
retina-face==0.0.17
customtkinter==5.2.2
rich==13.7.1
opencv-python==4.10.0.84
Pillow==10.4.0
